+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixed a couple of bugs in `SimpleFutureBuilder`.
 - Copied extension methods from [flutter_audio_games](https://pub.dev/packages/flutter_audio_games).
+- Added the `EnumListTile` widget.
 
 ## [0.26.0]